add Intro to sf R
The RConsortium-sponsored authors of SF have linked directly to OSGeo project trac and portal sites in a simple and accessible FAQ page. This minimal HTML page (and two logos) is both effective and light-weight. Please take a look at the page, I believe it will be a fine foundation for a lot of good OSGeo and RConsortium interaction ahead.